move store out in a file, add z index value
https://coveord.atlassian.net/browse/KIT-756
