+++ packages/atomic/src/components/facets-v1/facet-value-checkbox/facet-value-checkbox.pcss
@@ -5,11 +5,11 @@
   width: var(--atomic-facet-checkbox-size, 16px);
   height: var(--atomic-facet-checkbox-size, 16px);
 
-  @apply absolute z-10 left-2 flex justify-center rounded focus:outline-none focus:border-primary-light;
+  @apply absolute z-1 left-2 flex justify-center rounded focus:outline-none focus:border-primary-light;
 }
 
 .value-checkbox + label {
-  @apply w-full flex pl-8 pr-2 py-2.5 text-on-background cursor-pointer truncate hover:bg-neutral-light group;
+  @apply w-full flex pl-8 pr-2 py-2.5 text-on-background cursor-pointer truncate hover:bg-neutral-light;
 }
 
 /**
