+++ packages/atomic/src/utils/initialization-utils.tsx
@@ -1,32 +1,9 @@
-import {SearchEngine, NumericFacetValue, DateFacetValue} from '@coveo/headless';
+import {SearchEngine} from '@coveo/headless';
 import {ComponentInterface, getElement, h, forceUpdate} from '@stencil/core';
 import {i18n, TOptions} from 'i18next';
 import {ObservableMap} from '@stencil/store';
 import {buildCustomEvent} from './event-utils';
-
-interface FacetLabel {
-  label: string;
-}
-
-interface FacetValueFormat<ValueType> {
-  format(facetValue: ValueType): string;
-}
-
-type FacetStore<F extends FacetLabel> = Record<string, F>;
-
-export type AtomicStore = {
-  facets: FacetStore<FacetLabel>;
-  numericFacets: FacetStore<FacetLabel & FacetValueFormat<NumericFacetValue>>;
-  dateFacets: FacetStore<FacetLabel & FacetValueFormat<DateFacetValue>>;
-  categoryFacets: FacetStore<FacetLabel>;
-};
-
-export const initialStore: () => AtomicStore = () => ({
-  facets: {},
-  numericFacets: {},
-  dateFacets: {},
-  categoryFacets: {},
-});
+import {AtomicStore} from './store';
 
 /**
  * Bindings passed from the `AtomicSearchInterface` to its children components.
@@ -112,7 +89,9 @@ export function InitializeBindings() {
             this.bindings.i18n.off('languageChanged', updateLanguage);
 
           try {
-            this.initialize && this.initialize();
+            // When no controller is initialized, updating a property with a State() decorator, there will be no re-render.
+            // In this case, we have to manually trigger it.
+            this.initialize ? this.initialize() : forceUpdate(this);
           } catch (e) {
             this.error = e;
           }
